lifespan event moved to main.py
