+++ app/database/database.py
@@ -1,8 +1,6 @@
 from motor.motor_asyncio import AsyncIOMotorClient, AsyncIOMotorDatabase
 from app.config.settings import settings
 from app.logs.logging_config import database_logger
-from contextlib import asynccontextmanager
-from fastapi import FastAPI
 from pymongo.errors import OperationFailure
 # custom exceptions
 from app.exceptions.custom_exceptions import CreateConnectionException
@@ -66,16 +64,3 @@ class Database:
 
 
 motor_db = Database()
-
-
-@asynccontextmanager
-async def lifespan(_app: FastAPI):
-    """
-    Code before `yield`:
-        sets up database before actually starting taking requests
-    Code after `yield`:
-        cleans up the database right after the shutdown of the app
-    """
-    await motor_db.connect_and_init_db()
-    yield
-    await motor_db.close_db_connection()
